Removed legacy x86 build support
Moved to an exclusive x64 build project.
Updated base Windows support to Windows 10.
Added support to upload SRGB textures to OpenGL.
Update ffmpeg lib.
Small bug fixes.

+++ horseradish/src/common/meshAnim.hpp
@@ -151,7 +151,7 @@ namespace hr::geom
 				enum class Target{ None, Scale, Rotation, Translation };
 
 				Target target{ Target::None };
-				size_t jointIndex{ 0 };
+				uint32_t jointIndex{ 0 };
 				std::vector<Vector4f> frameData;
 			};
